+++ src/charm_state.py
@@ -71,6 +71,7 @@ COS_AGENT_INTEGRATION_NAME = "cos-agent"
 DEBUG_SSH_INTEGRATION_NAME = "debug-ssh"
 IMAGE_INTEGRATION_NAME = "image"
 MONGO_DB_INTEGRATION_NAME = "mongodb"
+PLANNER_INTEGRATION_NAME = "planner"
 
 LogLevel = Literal["CRITICAL", "FATAL", "ERROR", "WARNING", "INFO", "DEBUG"]
 
@@ -166,6 +167,48 @@ def _parse_labels(labels: str) -> tuple[str, ...]:
 
     return tuple(valid_labels)
 
+class PlannerConfig(BaseModel):
+    """Configuration for the planner service.
+    
+    Attributes:
+        token: Token for the planner service.
+        url: URL of the planner service.
+        flavor: The name of the flavor to register with the planner.
+    """
+
+    token: str
+    endpoint: AnyHttpUrl
+    flavor: str
+    
+    @classmethod
+    def from_charm(cls, charm: CharmBase) -> "PlannerConfig" | None:
+        """Initialize the config from charm.
+
+        Args:
+            charm: The charm instance.
+
+        Raises:
+            CharmConfigInvalidError: If an invalid configuration was set.
+
+        Returns:
+            Current planner config.
+        """
+        relations = charm.model.relations[PLANNER_INTEGRATION_NAME]
+        if not relations or not (relation := relations[0]).units:
+            return None
+        for unit in relation.units:
+            relation_data = relation.data[unit]
+            if not relation_data:
+                continue
+            token_secret_id = relation_data.get("token", None)
+            if token_secret_id is None:
+                return None
+            token_secret = charm.model.get_secret(token_secret_id)
+            return PlannerConfig(
+                token=token_secret.get_content()["token"],
+                endpoint=relation_data.get("endpoint", None),
+                flavor=relation.id,
+            )
 
 class RepoPolicyComplianceConfig(BaseModel):
     """Configuration for the repo policy compliance service.
@@ -252,6 +295,7 @@ class CharmConfig(BaseModel):
     aproxy_redirect_ports: list[str] = []
     custom_pre_job_script: str | None
     runner_manager_log_level: LogLevel
+    planner: PlannerConfig | None
 
     @classmethod
     def _parse_dockerhub_mirror(cls, charm: CharmBase) -> str | None:
@@ -523,6 +567,7 @@ class CharmConfig(BaseModel):
         runner_manager_log_level = cast(
             LogLevel, charm.config.get(RUNNER_MANAGER_LOG_LEVEL_CONFIG_NAME, "INFO")
         )
+        planner = PlannerConfig.from_charm(charm)
         # pydantic allows to pass str as AnyHttpUrl, mypy complains about it
         return cls(
             allow_external_contributor=cast(
@@ -546,6 +591,7 @@ class CharmConfig(BaseModel):
             ),
             custom_pre_job_script=custom_pre_job_script,
             runner_manager_log_level=runner_manager_log_level,
+            planner=planner
         )