+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 This changelog documents user-relevant changes to the GitHub runner charm.
 
+## 2025-01-30
+
+- Add support to integrate with GitHub Runner planner charm with the `github_runner_planner_v0` interface.
+
 ## 2025-01-14
 
 - Add support for running up to 100 multiple GitHub runner applications on a single instance.